Add ChatGPT 5 Pro support
Fixes problem where ChatDinger would sound one second after starting a ChatGPT 5 Pro prompt.

ChatDinger normally looks at the 'Execute' button (up arrow) which changes to a 'Stop' button for non-ChatGPT 5 Pro prompts while the prompt is being executed, then changes to the 'Use Voice Mode' button when the prompt finishes.

The problem is that for ChatGPT 5 Pro, the 'Execute' button immediately changes to the square 'Stop' button for 1 second, then changes to the 'Use Voice Mode' button and a new 'Stop' button appears in the 'Thinking' area. The change of the 'Stop' button to the 'Use Voice Mode' button in the input box after one second fools it into thinking that the prompt has completed.